python_sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentfriendly"
version = "0.1.0"
description = "Make your Python web app agent-friendly — detection, markdown serving, tool registry, x402 payments, multi-tenancy"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "AgentFriendly", email = "sdk@agentfriendly.dev" }]
keywords = ["ai", "agents", "llm", "fastapi", "django", "flask", "middleware"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Framework :: Django",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware",
    "Typing :: Typed",
]

# Core dependencies — all lightweight, no heavy ML libraries
dependencies = [
    "httpx>=0.27.0",          # HTTP client for key fetches (RFC 9421)
    "cryptography>=42.0.0",   # Ed25519 signature verification
    "pyjwt>=2.8.0",           # JWT signing/verification (Clawdentity + multi-tenancy)
    "fnmatch2>=0.0.7",        # Glob pattern matching for access control
]

[project.optional-dependencies]
# Content conversion (optional — pulls in heavier HTML parsing libraries)
content = [
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.12.0",
    "lxml>=5.0.0",
]

# Analytics storage adapters (optional)
analytics-sqlite = []  # uses stdlib sqlite3
analytics-postgres = ["asyncpg>=0.29.0"]
analytics-clickhouse = ["clickhouse-connect>=0.7.0"]

# Framework adapters (install via extras)
fastapi = ["fastapi>=0.111.0", "starlette>=0.37.0"]
django = ["django>=4.2"]
flask = ["flask>=3.0.0"]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "agentfriendly[content,fastapi,django,flask]",
]

[project.urls]
Homepage = "https://agentfriendly.dev"
Repository = "https://github.com/Jana-kabrit/AgentFriendly"
Documentation = "https://docs.agentfriendly.dev"

[tool.hatch.build.targets.wheel]
packages = ["agentfriendly"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["--tb=short", "-q"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
